Enable Gles2TexturePlayer to switch between flat and animated scenes
The Android frontend app now has a checkbox to select whether the
rendering is 'animated'.

A little more work would be need for Linux with a command line argument

Add shader program for a flat display when rendering.

shader_flat.c contains the vertices and vertex shader for rendering a
flat display. In practice this will stretch to the screen dimensions.

shader_anim.c contains the vertices and vertex shader for an animated
scene.

Also removed (apparently) redundant call to glDrawArrays().

Signed-off-by: Tony Lofthouse <tony.lofthouse@ti.com>